// src/models/SocietyVisitorVehicleEntriesModel.tsx
export interface SocietyVisitorVehicleEntriesModel {
    SocietyVisitorVehicleEntriesId: string;
    VehicleNumber: string;
    Company: string;
    DriverName: string;
    MobileNumber: string;
    Remarks: string;
    VisitType: string;
    VisitStatus: string;
    WelcomeCode: string;
    PlannedDateTime: Date;
    VisitorInDateTime: Date;
    VisitorOutDateTime: Date;
    VisitToUserId: string;
    SocietyId: string;
    VisitToSocietyBuildingUnitId: string;
    VisitToUserType: string;
    VisitToUserTypeUserId: string;

}